Breaking Change: Change default StoreGroup
- Make `CQRSStoreGroup` default
   - Rename `CQRSStoreGroup` to `StoreGroup`
- Remove old `StoreGroup` and `QueuedStoreGroup`
   -  Carve out this to module https://github.com/almin/legacy-store-group
