+++ src/UILayer/StoreGroupValidator.ts
@@ -1,9 +1,7 @@
 // LICENSE : MIT
 "use strict";
 const assert = require("assert");
-import { AnyStore, Store } from "../Store";
 import { Dispatcher } from "../Dispatcher";
-import { StoreLike } from "../StoreLike";
 /*
  StoreGroup
 
@@ -13,26 +11,12 @@ import { StoreLike } from "../StoreLike";
 
  */
 export class StoreGroupValidator {
-    /**
-     * validate stores in StoreGroup
-     * @param {Store[]} stores
-     */
-    static validateStores(stores: Array<AnyStore>): void | never {
-        stores.forEach(store => {
-            assert.ok(Store.isStore(store), `${store} should be instance of Store`);
-            assert.ok(typeof store.getState === "function", `${store} should implement getState() method.
-StoreGroup merge values of store*s*.`);
-            const storeName = store.name;
-            assert.ok(storeName, `${store} should have name property value.`);
-        });
-    }
 
     /**
      * validate the instance is StoreGroup-like object
-     * {@link Context} treat StoreGroup like object as StoreGroup.
-     * @param {*|StoreGroup|Store} storeGroup
+     * Context treat StoreGroup like object as StoreGroup.
      */
-    static validateInstance(storeGroup: any | StoreLike<any>): void | never {
+    static validateInstance(storeGroup: any): void | never {
         assert.ok(storeGroup !== undefined, "store should not be undefined");
         assert.ok(Dispatcher.isDispatcher(storeGroup), "storeGroup should inherit CoreEventEmitter");
         assert.ok(typeof storeGroup.onChange === "function", "StoreGroup should have #onChange method");
